Added preserving the framebuffer mode if required.
Added a switch to prevent restoring old framebuffer mode on exit.
Normally, the old framebuffer mode is restored when framebuffer is
closed on exit by FBIOPUT_VSCREENINFO ioctl(). Sometimes this
results in clearing the screen. When a frame is to be left on the
screen after exit, this is not desirable. Tested on imx28evk board
and PC.