A simpler implementation for 'run', sleep problem reading pipe.
run() now accepts a number of frames to display, so it is possible
to play a part or the whole animation with one number. It
simplifies the code and its usage.

When the interpret_commands() is waiting for new commands, fgets()
does not block but returns EOF on a pipe that has been closed by a
writer. This makes interpret_command() spin in a loop which loads
the CPU. The problem might be solved by closing and opening again
the pipe.